Address PR #941 review: split test files and add BT-debt TODO markers
Test split (AC-4, CS-18-004):
- Split test_create.py (807 lines) into three focused files:
  - test_create.py: CBT-05-001..004 bootstrap use-case tests only
  - test_bootstrap_participants.py: CBT-05-005/006 embedded-participant
    storage and M4 regression (#561/#562/#563)
  - test_helpers.py: CBT-05-006/007 _ensure_reporter_participant seeding
    and RM.ACCEPTED upgrade (#589/#624)
- Split test_update.py: extract BT tree-structure assertions into
  test_update_bt.py, keeping use-case behavior tests in test_update.py

BT-migration debt markers (BT-06-001, BT-15-001):
- AcceptInviteActorToCaseReceivedUseCase.execute(): replace comment
  presenting inline RM transitions as the architectural solution with an
  explicit TODO(#943) marking them as BT-migration debt
- _ensure_reporter_participant(): add TODO(#943) on direct
  ParticipantStatus(rm_state=RM.ACCEPTED) creation/upgrade
- test_helpers.py: add class-level TODO(#943) notes so the tests document
  the debt they cover
- Created tracking issue #943 for the BT migration work

3205 passed, 36 skipped in 42s
