+++ test/core/use_cases/received/case/test_update.py
@@ -10,7 +10,12 @@
 #  ("Third Party Software"). See LICENSE.md for more details.
 #  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
 #  U.S. Patent and Trademark Office by Carnegie Mellon University
-"""Tests for case-related use-case update handlers."""
+"""Tests for case-related use-case update handlers.
+
+BT structure tests (tree shape and no-post-BT-broadcast contract) have been
+extracted to ``test_update_bt.py`` to separate use-case behavior assertions
+from tree-structure assertions.
+"""
 
 import logging
 from typing import cast
@@ -20,15 +25,6 @@ from vultron.core.models.activity import VultronActivity
 from vultron.core.models.case_actor import VultronCaseActor
 from vultron.core.use_cases.received.case.update import (
     UpdateCaseReceivedUseCase,
-)
-from vultron.core.behaviors.case.update_tree import (
-    create_update_case_received_tree,
-)
-from vultron.core.behaviors.case.nodes.update import (
-    ApplyCaseUpdateNode,
-    BroadcastCaseUpdateNode,
-    CaptureCaseUpdateBroadcastExclusionsNode,
-    CheckCaseUpdateOwnerNode,
 )
 from vultron.wire.as2.rehydration import rehydrate as real_rehydrate
 from vultron.wire.as2.vocab.objects.case_participant import CaseParticipant
@@ -482,78 +478,3 @@ class TestCaseUseCases:
         broadcast = cast(VultronActivity, broadcast)
         assert broadcast.to is not None
         assert set(broadcast.to) == {alice, bob}
-
-    def test_update_case_bt_structure_includes_broadcast_node(
-        self, make_payload
-    ):
-        """UpdateCaseBT keeps ownership, embargo, update, and broadcast in-tree."""
-        owner_id = "https://example.org/users/owner"
-        case_id = "https://example.org/cases/bt1"
-        updated_case = VulnerabilityCase(
-            id_=case_id, name="Updated", attributed_to=owner_id
-        )
-        activity = update_case_activity(updated_case, actor=owner_id)
-        event = make_payload(activity)
-
-        tree = create_update_case_received_tree(
-            case_id=case_id,
-            actor_id=owner_id,
-            request=event,
-        )
-
-        assert tree.name == "UpdateCaseBT"
-        assert [child.__class__ for child in tree.children] == [
-            CheckCaseUpdateOwnerNode,
-            CaptureCaseUpdateBroadcastExclusionsNode,
-            ApplyCaseUpdateNode,
-            BroadcastCaseUpdateNode,
-        ]
-
-    def test_update_case_bt_executes_without_post_bt_broadcast(
-        self, make_payload, monkeypatch
-    ):
-        """UpdateCaseBT handles the broadcast internally instead of after execute()."""
-        dl = SqliteDataLayer("sqlite:///:memory:")
-        owner_id = "https://example.org/users/owner"
-        participant_id = "https://example.org/users/alice"
-        case_id = "https://example.org/cases/bt2"
-
-        case_actor = VultronCaseActor(
-            id_=f"{case_id}/actor",
-            name=f"CaseActor for {case_id}",
-            attributed_to=owner_id,
-            context=case_id,
-        )
-        dl.create(case_actor)
-
-        case = VulnerabilityCase(
-            id_=case_id,
-            name="Original",
-            attributed_to=owner_id,
-        )
-        case.actor_participant_index[participant_id] = (
-            "https://example.org/participants/p-bt2"
-        )
-        dl.create(case)
-
-        updated_case = VulnerabilityCase(
-            id_=case_id,
-            name="Updated",
-            attributed_to=owner_id,
-        )
-        activity = update_case_activity(updated_case, actor=owner_id)
-        event = make_payload(activity)
-
-        def _should_not_be_called(*args, **kwargs):
-            raise AssertionError("post-BT broadcast helper should not run")
-
-        monkeypatch.setattr(
-            UpdateCaseReceivedUseCase,
-            "_broadcast_case_update",
-            _should_not_be_called,
-        )
-
-        UpdateCaseReceivedUseCase(dl, event).execute()
-
-        outbox_items = dl.outbox_list_for_actor(case_actor.id_)
-        assert len(outbox_items) == 1
